Slight refactor to auth context, fixes lint warning
