+++ frontend/src/context/useAuth.ts
@@ -1,0 +1,8 @@
+import { AuthContext } from './AuthContext';
+import { useContext } from 'react';
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) throw new Error('useAuth must be used within an AuthProvider');
+  return context;
+};
